+++ matplotlive/__init__.py
@@ -7,13 +7,13 @@
 """Stream live plots to a matplotlib figure."""
 
 from .exceptions import MatplotliveError
-from .recent_past import RecentPast
+from .live_plot import LivePlot
 from .sketch import Sketch
 
 __version__ = "0.0.1"
 
 __all__ = [
     "MatplotliveError",
-    "RecentPast",
+    "LivePlot",
     "Sketch",
 ]